# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JudoClub Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clubs:
    get:
      operationId: listClubs
      tags: [clubs]
      summary: List all clubs
      responses:
        "200":
          description: List of clubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Club"
    post:
      operationId: createClub
      tags: [clubs]
      summary: Create a new club
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClubInput"
      responses:
        "201":
          description: Created club
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
  /clubs/{clubId}:
    get:
      operationId: getClub
      tags: [clubs]
      summary: Get club by ID
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
    patch:
      operationId: updateClub
      tags: [clubs]
      summary: Update club
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClubUpdate"
      responses:
        "200":
          description: Updated club
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
    delete:
      operationId: deleteClub
      tags: [clubs]
      summary: Delete club
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clubs/{clubId}/stats:
    get:
      operationId: getClubStats
      tags: [clubs]
      summary: Get club statistics
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubStats"
  /clubs/{clubId}/members:
    get:
      operationId: listClubMembers
      tags: [clubs]
      summary: List coaches/members of a club
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClubMember"
    post:
      operationId: addClubMember
      tags: [clubs]
      summary: Add a coach/member to a club
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClubMemberInput"
      responses:
        "201":
          description: Added member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubMember"
  /clubs/{clubId}/members/{memberId}:
    delete:
      operationId: removeClubMember
      tags: [clubs]
      summary: Remove a member from a club
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
  /clubs/{clubId}/payment-summary:
    get:
      operationId: getClubPaymentSummary
      tags: [payments]
      summary: Get payment summary for a club
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSummary"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [super_admin, club_admin, coach, athlete, parent]
        clubId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - email
        - role
        - createdAt
...
    Club:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        memberCount:
          type: integer
        athleteCount:
          type: integer
      required:
        - id
        - name
        - createdAt
        - memberCount
        - athleteCount
    ClubInput:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        country:
          type: string
        description:
          type: string
        logoUrl:
          type: string
      required:
        - name
    ClubUpdate:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        country:
          type: string
        description:
          type: string
        logoUrl:
          type: string
    ClubStats:
      type: object
      properties:
        clubId:
          type: integer
        totalAthletes:
          type: integer
        totalCoaches:
          type: integer
        totalCompetitions:
          type: integer
        totalFights:
          type: integer
        pendingPayments:
          type: integer
        paidPayments:
          type: integer
        overduePayments:
          type: integer
        recentCompetitions:
          type: array
          items:
            $ref: "#/components/schemas/Competition"
      required:
        - clubId
        - totalAthletes
        - totalCoaches
        - totalCompetitions
        - totalFights
        - pendingPayments
        - paidPayments
        - overduePayments
        - recentCompetitions
    ClubMember:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        clubId:
          type: integer
        role:
          type: string
          enum: [admin, coach]
        user:
          $ref: "#/components/schemas/User"
        joinedAt:
          type: string
      required:
        - id
        - userId
        - clubId
        - role
        - joinedAt
    ClubMemberInput:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
          enum: [admin, coach]
      required:
        - email
        - role
...
    PaymentSummary:
      type: object
      properties:
        clubId:
          type: integer
        totalAmount:
          type: number
        paidAmount:
          type: number
        pendingAmount:
          type: number
        overdueAmount:
          type: number
        paidCount:
          type: integer
        pendingCount:
          type: integer
        overdueCount:
          type: integer
      required:
        - clubId
        - totalAmount
        - paidAmount
        - pendingAmount
        - overdueAmount
        - paidCount
        - pendingCount
        - overdueCount
    Competition:
      type: object
      properties:
        id:
          type: integer
        clubId:
          type: integer
        name:
          type: string
        location:
          type: ["string", "null"]
        date:
          type: string
        format:
          type: string
          enum: [olympic, round_robin]
        fightDurationSeconds:
          type: integer
        status:
          type: string
          enum: [draft, active, finished]
        participantCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - clubId
        - name
        - date
        - format
        - fightDurationSeconds
        - status
        - participantCount
        - createdAt
...