# lib/api-zod/src/generated/types/paymentSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JudoClub Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentSummary {
  clubId: number;
  totalAmount: number;
  paidAmount: number;
  pendingAmount: number;
  overdueAmount: number;
  paidCount: number;
  pendingCount: number;
  overdueCount: number;
}